+++ ui/components/ui/sidebar/menu.tsx
@@ -113,6 +113,7 @@ export const Menu = ({ isOpen }: { isOpen: boolean }) => {
                           target={menu.target}
                           tooltip={menu.tooltip}
                           isOpen={isOpen}
+                          highlight={menu.highlight}
                         />
                       )}
                     </div>